requirements.txt
tensorflow-cpu
opencv-python
numpy